pyproject.toml
[project]
name = "boundbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.35.3",
    "openai>=2.1.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "vllm>=0.6.0",
    "pyyaml",
]
